a function that returns the length of a string

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -11,7 +11,7 @@
  * Return: The string to check the length of
  */
 
-int _strlen(const char *s)
+size_t _strlen(const char *s)
 {
 	size_t length = 0;
